colocate buffer commands
