+++ src/commands/brpoplpush.js
@@ -1,3 +1,8 @@
 export function brpoplpush(source, destination) {
   return this.rpoplpush(source, destination)
 }
+
+export function brpoplpushBuffer(source, destination) {
+  const valP = brpoplpush.apply(this, [source, destination])
+  return valP.then(val => (val ? Buffer.from(val) : val))
+}
