+++ src/commands/hgetall.js
@@ -1,3 +1,12 @@
 export function hgetall(key) {
   return this.data.get(key) || {}
 }
+
+export function hgetallBuffer(key) {
+  const val = hgetall.apply(this, [key])
+  Object.keys(val).forEach(keyInObject => {
+    val[keyInObject] = Buffer.from(val[keyInObject])
+  })
+
+  return val
+}
