+++ src/commands/hmget.js
@@ -7,3 +7,8 @@ export function hmget(key, ...fields) {
     return hash[field]
   })
 }
+
+export function hmgetBuffer(key, ...fields) {
+  const val = hmget.apply(this, [key, ...fields])
+  return val.map(payload => (payload ? Buffer.from(payload) : payload))
+}
