--- src/commands/hmgetBuffer.js
@@ -1,6 +1,0 @@
-import { hmget } from './hmget'
-
-export function hmgetBuffer(key, ...fields) {
-  const val = hmget.apply(this, [key, ...fields])
-  return val.map(payload => (payload ? Buffer.from(payload) : payload))
-}
