+++ src/commands/index.js
@@ -3,7 +3,6 @@ export * from './auth'
 export * from './bgrewriteaof'
 export * from './bgsave'
 export * from './brpoplpush'
-export * from './brpoplpushBuffer'
 export * from './connect'
 export * from './dbsize'
 export * from './decr'
@@ -27,14 +26,11 @@ export * from './hdel'
 export * from './hexists'
 export * from './hget'
 export * from './hgetall'
-export * from './hgetallBuffer'
-export * from './hgetBuffer'
 export * from './hincrby'
 export * from './hincrbyfloat'
 export * from './hkeys'
 export * from './hlen'
 export * from './hmget'
-export * from './hmgetBuffer'
 export * from './hmset'
 export * from './hscan'
 export * from './hset'
@@ -51,7 +47,6 @@ export * from './lindex'
 export * from './linsert'
 export * from './llen'
 export * from './lpop'
-export * from './lpopBuffer'
 export * from './lpush'
 export * from './lpushx'
 export * from './lrange'
@@ -76,9 +71,7 @@ export * from './rename'
 export * from './renamenx'
 export * from './role'
 export * from './rpop'
-export * from './rpopBuffer'
 export * from './rpoplpush'
-export * from './rpoplpushBuffer'
 export * from './rpush'
 export * from './rpushx'
 export * from './sadd'
