+++ src/commands/lpop.js
@@ -10,3 +10,8 @@ export function lpop(key) {
 
   return item
 }
+
+export function lpopBuffer(key) {
+  const val = lpop.apply(this, [key])
+  return val ? Buffer.from(val) : val
+}
