--- src/commands/rpopBuffer.js
@@ -1,5 +1,0 @@
-import { rpop } from './rpop'
-
-export function rpopBuffer(key) {
-  return rpop.apply(this, [key])
-}
