--- src/commands/rpoplpushBuffer.js
@@ -1,5 +1,0 @@
-import { rpoplpush } from './rpoplpush'
-
-export function rpoplpushBuffer(source, destination) {
-  return rpoplpush.apply(this, [source, destination])
-}
